#8 [MOD] 홈 -> 메인으로 수정
