+++ src/pages/main/components/CategorySection.tsx
