+++ src/pages/main/components/FriendHeader.tsx
