+++ src/pages/main/components/Streak.tsx
