+++ src/routes/homeRoutes.tsx
@@ -1,10 +1,12 @@
-import type {RouteType} from "../types/routeType.ts";
+import type { RouteType } from "../types/routeType.ts";
 import routePath from "./routePath.ts";
 import HomePage from "../pages/home/HomePage.tsx";
 
-const homeRoutes: RouteType[] = [{
+const homeRoutes: RouteType[] = [
+  {
     path: routePath.HOMEPAGE,
     element: <HomePage />,
-}]
+  },
+];
 
 export default homeRoutes;
