+++ src/routes/pageRoutes.tsx
@@ -9,10 +9,12 @@ import alarmRoutes from "./alarmRoutes.tsx";
 import friendRoutes from "./friendRoutes.tsx";
 import recordRoutes from "./recordRoutes.tsx";
 import quizRoutes from "./quizRoutes.tsx";
+import mainRoutes from "./mainRoutes.tsx";
 
 const allRoutes = [
   ...exampleRoutes,
   ...homeRoutes,
+  ...mainRoutes,
   ...recordRoutes,
   ...quizRoutes,
   ...myPageRoutes,
